Remove glut/freeglut and add GLFW

+++ src/Scene/MainMenu/mainMenu.c
@@ -78,7 +78,7 @@ int MainMenu_keyUp(InputType inputType) {
     return 0;
 }
 
-int MainMenu_mouseMovement(float x, float y) {
+int MainMenu_mouseMovement(double x, double y) {
     Camera *cam = &StateManager_top(&engine.sM)->camera;
     // If cursor is locked, let the camera move, else ignore movement
     if (engine.lockCamera) {
